F_T30603 move MapBundle Value Objects to core
- adjust use statements
