+++ demosplan/DemosPlanCoreBundle/Logic/Map/MapImageToPolygonCropper.php
@@ -12,9 +12,9 @@ declare(strict_types=1);
 
 namespace demosplan\DemosPlanCoreBundle\Logic\Map;
 
-use demosplan\DemosPlanMapBundle\ValueObject\CoordinatesViewport;
-use demosplan\DemosPlanMapBundle\ValueObject\Feature;
-use demosplan\DemosPlanMapBundle\ValueObject\MapLayer;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\CoordinatesViewport;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\Feature;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\MapLayer;
 use Symfony\Component\Filesystem\Filesystem;
 use Tightenco\Collect\Support\Collection;
