+++ demosplan/DemosPlanCoreBundle/Logic/Map/MapLayerMerger.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 
 namespace demosplan\DemosPlanCoreBundle\Logic\Map;
 
-use demosplan\DemosPlanMapBundle\ValueObject\MapLayer;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\MapLayer;
 use Exception;
 use Symfony\Component\Filesystem\Filesystem;
