+++ demosplan/DemosPlanCoreBundle/Transformers/Map/MapOptionsTransformer.php
@@ -11,7 +11,7 @@
 namespace demosplan\DemosPlanCoreBundle\Transformers\Map;
 
 use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\Transformer\BaseTransformer;
-use demosplan\DemosPlanMapBundle\ValueObject\MapOptions;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\MapOptions;
 use TypeError;
 
 class MapOptionsTransformer extends BaseTransformer
