+++ tests/backend/core/Map/Unit/PrintLayerToMapLayerConverterTest.php
@@ -14,10 +14,10 @@ namespace Tests\Core\Map\Unit;
 
 use demosplan\DemosPlanCoreBundle\Logic\Map\PrintLayerToMapLayerConverter;
 use demosplan\DemosPlanCoreBundle\Utilities\DemosPlanPath;
-use demosplan\DemosPlanMapBundle\ValueObject\CoordinatesViewport;
-use demosplan\DemosPlanMapBundle\ValueObject\PrintLayer;
-use demosplan\DemosPlanMapBundle\ValueObject\PrintLayerTile;
-use demosplan\DemosPlanMapBundle\ValueObject\PrintLayerTilePosition;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\CoordinatesViewport;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\PrintLayer;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\PrintLayerTile;
+use demosplan\DemosPlanCoreBundle\ValueObject\Map\PrintLayerTilePosition;
 use Intervention\Image\ImageManager;
 use Tests\Base\UnitTestCase;
 use Tightenco\Collect\Support\Collection;
